LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/path-objects/2d.path.rect.nonfinite-expected.txt

2d.path.rect.nonfinite
rect() with Infinity/NaN is ignored
Actual output:
Expected output:


PASS rect() with Infinity/NaN is ignored 
